test: split absolute directory regressions
